+++ CHANGES.rst
@@ -2,6 +2,16 @@
 Changelog
 ---------
 
+Version 0.2
+-----------
+
+Released on 2012-07-26.
+
+- Fixed a bug related to update command (on Windows)
+- Used UTC datetime
+- Decoded pakbus packet properly
+- Other minor bugs fixed
+
 Version 0.1
 -----------